- made http status code 401 response compliant (WWW-Authenticate) header - added new conditions authorizationexception to help 401 - updated user sample showing header parameters and header overrides - removed the 'X-' in front of custom headers - added some helpful logging

+++ samples/sample_user_service/src/com/tales/samples/userservice/UserResource.java
@@ -22,7 +22,7 @@ import java.util.UUID;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
-
+import com.tales.contracts.services.http.HeaderParam;
 import com.tales.contracts.services.http.PathParam;
 import com.tales.contracts.services.http.RequestParam;
 import com.tales.contracts.services.http.ResourceContract;
@@ -68,22 +68,24 @@ public class UserResource {
 	 * The HTTP request to get a particular user.
 	 */
 	@ResourceOperation( name="get_user", path="GET : users/{id}" )
-	public TransportUser getUser(  @PathParam( name="id" )UUID theId  ) {
-		// Conditions.checkParameter throws InvalidParameterException which tales turns into HTTP status code 400
-		Conditions.checkParameter( theId != null , "id", "an id must be given" );
+	public TransportUser getUser( 
+			@PathParam( name="id" )UUID theId, 
+			@HeaderParam( name="Authorization" )String theAuthToken ) {
+		Conditions.checkAuthorization( validateAuthorization( theAuthToken ), "Sample", "unauthorized attempt to access resource" ); // if invalid, throws AuthorizationException which turns into HTTP status code 401
+		Conditions.checkParameter( theId != null , "id", "an id must be given" ); // if invalid, throws InvalidParameterException, which turns into HTTP status code 400
 		
 		User user = engine.getUser( theId );
-		// Conditions.checkFound throws a NotFoundException which tales turns into an HTTP status code 404
 		// public interface doesn't support sending soft-deleted entities
-		Conditions.checkFound( user != null && !user.isDeleted(), "Could not find the user with id '%s'.", theId ); 
+		Conditions.checkFound( user != null && !user.isDeleted(), "Could not find the user with id '%s'.", theId );  // if invalid, throws NotFoundException, which turns into HTTP status code 404
 		return TransportUser.toTransportUser( user );
 	}
 	
 	/**
 	 * The HTTP request to get all users.
 	 */
-	@ResourceOperation( name="get_users", path="GET : users" )
-	public Collection<TransportUser> getUsers( ) { // TODO: generally you don't want to get a list of them all, but allow filters or limiting how many, continuation tokens, etc
+	@ResourceOperation( name="get_users", path="GET : users"  )
+	public Collection<TransportUser> getUsers( @HeaderParam( name="Authorization" )String theAuthToken ) { // TODO: generally you don't want to get a list of them all, but allow filters or limiting how many, continuation tokens, etc
+		Conditions.checkAuthorization( validateAuthorization( theAuthToken ), "Sample", "unauthorized attempt to access resource" ); 
 		Collection<User> users = engine.getUsers( );
 		List<TransportUser> transportUsers = new ArrayList<TransportUser>(users.size());
 		
@@ -102,8 +104,9 @@ public class UserResource {
 	@ResourceOperation( name="create_user", path="GET | POST : users/create" ) // supporting GET just so a browser can easily be used for manual testing
 	public TransportUser createUser( 
 			 @RequestParam( name="first_name" )String theFirstName, 
-			 @RequestParam( name="last_name" )String theLastName ) {
-		// Conditions.checkParameter throws InvalidParameterException which tales turns into HTTP status code 400
+			 @RequestParam( name="last_name" )String theLastName,
+			 @HeaderParam( name="Authorization" )String theAuthToken ) {
+		Conditions.checkAuthorization( validateAuthorization( theAuthToken ), "Sample", "unauthorized attempt to access resource" ); 
 		Conditions.checkParameter( !Strings.isNullOrEmpty( theFirstName ), "first_name", "first name must be provided" );
 		Conditions.checkParameter( !Strings.isNullOrEmpty( theLastName ), "last_name", "last name must be provided" );
 		
@@ -118,14 +121,14 @@ public class UserResource {
 	@ResourceOperation( name="update_user", path="GET | POST : users/{id}/update" ) // supporting GET just so a browser can easily be used for manual testing
 	public TransportUser updateUser( 
 			@PathParam( name="id" )UUID theId, 
-			@RequestParam( name="user" )TransportUser theUser ) {
-		// Conditions.checkParameter throws InvalidParameterException which tales turns into HTTP status code 400
+			@RequestParam( name="user" )TransportUser theUser,
+			@HeaderParam( name="Authorization" )String theAuthToken ) {
+		Conditions.checkAuthorization( validateAuthorization( theAuthToken ), "Sample", "unauthorized attempt to access resource" ); 
 		Conditions.checkParameter( theId != null, "id", "an id must be given" );
 		Conditions.checkParameter( theUser != null, "user", "a user must be given" );
 		Conditions.checkParameter( theId.equals( theUser.getId( ) ), "id", "path id '%s' does not match the given user id '%s'", theId, theUser.getId() );
 		
 		User user = engine.updateUser( TransportUser.toEngineUser( theUser ) );
-		// Conditions.checkFound throws a NotFoundException which tales turns into an HTTP status code 404
 		Conditions.checkFound( user != null, "Could not find the user with id '%s'.", theId );
 		return TransportUser.toTransportUser( user );
 	}
@@ -134,12 +137,27 @@ public class UserResource {
 	 * The HTTP request to delete a user.
 	 */
 	@ResourceOperation( name="delete_user", path="GET | POST : users/{id}/delete" ) // supporting GET just so a browser can easily be used for manual testing
-	public void deleteUser( @PathParam( name="id" )UUID theId ) {
-		// Conditions.checkParameter throws InvalidParameterException which tales turns into HTTP status code 400
+	public void deleteUser(
+			@PathParam( name="id" )UUID theId, 
+			@HeaderParam( name="Authorization" )String theAuthToken ) {
+		Conditions.checkAuthorization( validateAuthorization( theAuthToken ), "Sample", "unauthorized attempt to access resource" ); 
 		Conditions.checkParameter( theId != null, "id", "an id must be given" );
 		
 		boolean deleted = engine.deleteUser( theId );
-		// Conditions.checkFound throws a NotFoundException which tales turns into an HTTP status code 404
 		Conditions.checkFound( deleted, "Could not find the user with id '%s'.", theId );
+	}
+	
+
+	/**
+	 * Validate the HTTP based auth header.
+	 * @param theAuthorization the string to very
+	 * @return th
+	 */
+	private boolean validateAuthorization( String theAuthorization ) {
+		if( !Strings.isNullOrEmpty( theAuthorization ) ) {
+			return theAuthorization.startsWith( "Sample" ); 
+		} else {
+			return false;
+		}
 	}
 }
